# splinterd/api/static/openapi.yml
...
openapi: '3.0.0'
info:
  version: 0.3.10
  title: SplinterD API
  description: REST API for SplinterD
servers:
  - url: http://localhost:9000/api
paths:
...
  /keys:
    get:
      tags:
        - Key Registry
      description: List public key information in the Key Registry
      parameters:
        - name: offset
          in: query
          description: paging offset
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          description: maximum number of items to return (max 100)
          required: false
          schema:
            type: integer
            default: 100
      responses:
        200:
          description: list of public key data objects
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PublicKeyInfo'
                  paging:
                    $ref: '#/components/schemas/Paging'
        400:
          description: Request was malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /keys/{public_key}:
    get:
      tags:
        - Key Registry
      description: Fetch public key information in the Key Registry by public key
      parameters:
        - name: public_key
          in: path
          description: public key to query, in hex
          required: true
          schema:
            type: string
      responses:
        200:
          description: Node
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/PublicKeyInfo"
        400:
          description: "{public_key} was malformed"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: The information for {public_key} was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /scabbard/{circuit}/{service_id}/batches:
    post:
      description: Send a list of Sabre batches to the specified Scabbard service
      parameters:
        - name: circuit
          in: path
          description: circuit the targeted service belongs to
          required: true
          schema:
            type: string
        - name: service_id
          in: path
          description: ID of the targeted service
          required: true
          schema:
            type: string
      responses:
        202:
          description: Batch has been submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        404:
          description: circuit id or service id not found
        500:
          description: Internal service error
  /scabbard/{circuit}/{service_id}/batch_statuses:
    post:
      description: Send a list of Sabre batches to the specified Scabbard service
      parameters:
        - name: circuit
          in: path
          description: circuit the targeted service belongs to
          required: true
          schema:
            type: string
        - name: service_id
          in: path
          description: ID of the targeted service
          required: true
          schema:
            type: string
        - name: ids
          in: query
          description: Comma separated list of batch ids
          required: true
          schema:
            type: string
        - name: wait
          in: query
          description: time (in seconds) to wait for batches to be committed
          required: false
          schema:
            type: integer
            default: 300
      responses:
        200:
          description: Batch has been submitted
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BatchStatus'
        404:
          description: circuit id or service id not found
        400:
          description: Malformed ids list
        500:
          description: Internal service error
components:
  schemas:
    Error:
      additionalProperties: false
      properties:
        message:
          description: A message describing the error that occurred
          type: string
          example: DatabaseError({description})
      required:
        - message
...
    PublicKeyInfo:
      type: object
      properties:
        public_key:
          type: string
        node_id:
          type: string
        metadata:
          type: object
          additionalProperties: true
          example:
            name: Jane User
            organization: Acme Corporation
    Link:
      type: object
      properties:
        link:
          type: string
          description: link to batch_statuses endpoint
          example: "/scabbard/{circuit_id}/{service_id}/batch_statuses?ids={ids}"
    BatchStatus:
      type: object
      properties:
        id:
          type: string
          description: Batch header signature
          example: 6ff35474a572087e08fd6a54d563bd8172951b363e5c9731f1a40a855e14bba45dac515364a08d8403f4fb5d4a206174b7f63c29e4f4e425dc71b95494b8a798
        status:
          type: object
          description: Batch status
          properties:
            statusType:
              type: string
              enum:
                - Unknown
                - Invalid
                - Valid
                - Pending
                - Committed
            message:
                type: array
                items:
                  type: object
                  properties:
                    transaction_id:
                      type: string
                      example: f4e147ff464013deccb3e68bb8619beffb29ff86b401257c93bcf8ef76d7ca173fa84b4f4a58414ad2d00a2c9f810cbb726e01cd26ebd44720239d9d35853099
                    error_message:
                      type: string
                      example: "Wasm contract returned invalid transaction: xo, 0.3.3"
                    error_data:
                      type: array
                      items:
                        type: integer
...
    Paging:
      type: object
      properties:
        current:
          type: string
        offset:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        first:
          type: string
        prev:
          type: string
        next:
          type: string
        last:
          type: string
      example:
        current: /nodes?offset=10&limit=10
        offset: 10
        limit: 10
        total: 50
        first: /nodes?offset=0&limit=10
        prev: /nodes?offset=0&limit=10
        next: /nodes?offset=20&limit=10
        last: /nodes?offset=40&limit=10